refactor(whatsrust): extract outbound message wrappers
